// src/__generated__/relayItemsDevices_query.graphql.ts
export type DeviceStatus = "ACTIVE" | "ARCHIVED" | "DECOMMISSIONED" | "DELETED" | "INACTIVE" | "MAINTENANCE" | "OFFLINE" | "ONLINE" | "PENDING" | "PENDING_DELETION" | "%future added value";
import type { FragmentRefs } from "relay-runtime";
export type relayItemsDevices_query$data = {
  readonly devices: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly displayName: string | null | undefined;
        readonly hostname: string | null | undefined;
        readonly id: string;
        readonly machineId: string;
        readonly nickname: string | null | undefined;
        readonly status: DeviceStatus | null | undefined;
      };
    }>;
  };
  readonly " $fragmentType": "relayItemsDevices_query";
};
export type relayItemsDevices_query$key = {
  readonly " $data"?: relayItemsDevices_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"relayItemsDevices_query">;
};