// src/__generated__/relayItemsOrgs_query.graphql.ts
export type relayItemsOrgs_query$data = {
  readonly organizations: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly category: string | null | undefined;
        readonly id: string;
        readonly name: string;
      };
    }>;
  };
  readonly " $fragmentType": "relayItemsOrgs_query";
};
export type relayItemsOrgs_query$key = {
  readonly " $data"?: relayItemsOrgs_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"relayItemsOrgs_query">;
};