// src/__generated__/relayItemsOrgsPaginationQuery.graphql.ts
export type relayItemsOrgsPaginationQuery$variables = {
  after?: string | null | undefined;
  first?: number | null | undefined;
  search?: string | null | undefined;
};
export type relayItemsOrgsPaginationQuery$data = {
  readonly " $fragmentSpreads": FragmentRefs<"relayItemsOrgs_query">;
};
export type relayItemsOrgsPaginationQuery = {
  response: relayItemsOrgsPaginationQuery$data;
  variables: relayItemsOrgsPaginationQuery$variables;
};