// src/__generated__/relayItemsSoftware_query.graphql.ts
export type relayItemsSoftware_query$data = {
  readonly softwares: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly currentVersion: string | null | undefined;
        readonly id: string;
        readonly name: string;
        readonly publisher: string | null | undefined;
      };
    }>;
  };
  readonly " $fragmentType": "relayItemsSoftware_query";
};
export type relayItemsSoftware_query$key = {
  readonly " $data"?: relayItemsSoftware_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"relayItemsSoftware_query">;
};